requirements.txt
openai-whisper
transformers
pytest
fastapi
uvicorn[standard]
python-multipart
jinja2
